backend/requirements.txt
asgiref==3.5.0
sqlparse==0.4.2
tzdata==2021.5
psycopg2-binary>=2.8
mysqlclient>=2.0
black==22.1.0
djangorestframework==3.13.1
django-cors-headers==3.11.0
drf-yasg==1.20.0